Centralize connectivity checks
