+++ core/src/main/java/com/uoa/core/network/NetworkMonitorImpl.kt
@@ -2,36 +2,76 @@ package com.uoa.core.network
 
 import android.content.Context
 import android.net.ConnectivityManager
+import android.net.ConnectivityManager.NetworkCallback
 import android.net.Network
 import android.net.NetworkCapabilities
-import android.net.NetworkRequest
-import com.uoa.core.utils.internetconnectivity.NetworkMonitor
+import android.net.NetworkRequest.Builder
+import androidx.compose.ui.util.trace
+import androidx.core.content.getSystemService
+import dagger.hilt.android.qualifiers.ApplicationContext
+import com.uoa.core.network.DaDispatchers
+import com.uoa.core.network.Dispatcher
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.conflate
+import kotlinx.coroutines.flow.flowOn
+import javax.inject.Inject
 
-class NetworkMonitorImpl(private val context: Context) : NetworkMonitor {
+class NetworkMonitorImpl @Inject constructor(
+    @ApplicationContext private val context: Context,
+    @Dispatcher(DaDispatchers.IO) private val ioDispatcher: CoroutineDispatcher,
+) : NetworkMonitor {
 
     override val isOnline: Flow<Boolean> = callbackFlow {
-        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        val networkCallback = object : ConnectivityManager.NetworkCallback() {
-            override fun onAvailable(network: Network) {
-                trySend(true).isSuccess
+        trace("NetworkMonitor.callbackFlow") {
+            val connectivityManager = context.getSystemService<ConnectivityManager>()
+            if (connectivityManager == null) {
+                trySend(false)
+                close()
+                return@callbackFlow
             }
 
-            override fun onLost(network: Network) {
-                trySend(false).isSuccess
+            val callback = object : NetworkCallback() {
+
+                private val networks = mutableSetOf<Network>()
+
+                override fun onAvailable(network: Network) {
+                    networks += network
+                    trySend(true)
+                }
+
+                override fun onLost(network: Network) {
+                    networks -= network
+                    trySend(networks.isNotEmpty())
+                }
             }
-        }
 
-        val networkRequest = NetworkRequest.Builder()
-            .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-            .build()
+            trace("NetworkMonitor.registerNetworkCallback") {
+                val request = Builder()
+                    .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+                    .build()
+                connectivityManager.registerNetworkCallback(request, callback)
+            }
 
-        connectivityManager.registerNetworkCallback(networkRequest, networkCallback)
+            trySend(connectivityManager.isCurrentlyConnected())
 
-        awaitClose {
-            connectivityManager.unregisterNetworkCallback(networkCallback)
+            awaitClose {
+                connectivityManager.unregisterNetworkCallback(callback)
+            }
         }
     }
+        .flowOn(ioDispatcher)
+        .conflate()
+
+    @Suppress("DEPRECATION")
+    private fun ConnectivityManager.isCurrentlyConnected() = when {
+        true ->
+            activeNetwork
+                ?.let(::getNetworkCapabilities)
+                ?.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+
+        else -> activeNetworkInfo?.isConnected
+    } ?: false
 }
